Rename threads to thread to match the Rust standard library.
