+++ src/lib.rs
@@ -27,7 +27,7 @@ mod operator;
 mod range_map;
 mod shims;
 mod stacked_borrows;
-mod threads;
+mod thread;
 
 // Make all those symbols available in the same place as our own.
 pub use rustc_mir::interpret::*;
@@ -42,7 +42,7 @@ pub use crate::shims::intrinsics::EvalContextExt as IntrinsicsEvalContextExt;
 pub use crate::shims::os_str::EvalContextExt as OsStrEvalContextExt;
 pub use crate::shims::panic::{CatchUnwindData, EvalContextExt as PanicEvalContextExt};
 pub use crate::shims::sync::{EvalContextExt as SyncEvalContextExt};
-pub use crate::shims::threads::EvalContextExt as ThreadShimsEvalContextExt;
+pub use crate::shims::thread::EvalContextExt as ThreadShimsEvalContextExt;
 pub use crate::shims::time::EvalContextExt as TimeEvalContextExt;
 pub use crate::shims::tls::{EvalContextExt as TlsEvalContextExt, TlsData};
 pub use crate::shims::EvalContextExt as ShimsEvalContextExt;
@@ -63,7 +63,7 @@ pub use crate::range_map::RangeMap;
 pub use crate::stacked_borrows::{
     EvalContextExt as StackedBorEvalContextExt, Item, Permission, PtrId, Stack, Stacks, Tag,
 };
-pub use crate::threads::{
+pub use crate::thread::{
     EvalContextExt as ThreadsEvalContextExt, ThreadId, ThreadManager, ThreadState,
 };
